show: to avoid strange effects with --scale, mouseNav and deltaMouse variables must be floating point variables
git-svn-id: svn+ssh://svn.code.sf.net/p/slam6d/code/trunk@2081 51f1e6e7-7289-49ad-8748-f09e43fc6daa

+++ src/show/show_common.cc
@@ -137,7 +137,7 @@ int show_poses           = 1;            // Show the coordinate axes of the pose
  */
 int cameraNavMouseMode  = 1;
 
-int mouseNavX, mouseNavY;
+double mouseNavX, mouseNavY;
 int mouseNavButton = -1;
 int mousePresX, mousePresY;
